requirements.txt
django==6.0.3
djangorestframework==3.17.1
openpyxl==3.1.5
